Add SolanaTransactionDecoder class
