+++ rotkehlchen/rotkehlchen.py
@@ -48,6 +48,7 @@ from rotkehlchen.chain.polygon_pos.node_inquirer import PolygonPOSInquirer
 from rotkehlchen.chain.scroll.manager import ScrollManager
 from rotkehlchen.chain.scroll.node_inquirer import ScrollInquirer
 from rotkehlchen.chain.solana.manager import SolanaManager
+from rotkehlchen.chain.solana.node_inquirer import SolanaInquirer
 from rotkehlchen.chain.substrate.manager import SubstrateManager
 from rotkehlchen.chain.substrate.utils import (
     KUSAMA_NODES_TO_CONNECT_AT_START,
@@ -472,8 +473,11 @@ class Rotkehlchen:
             bitcoin_manager=BitcoinManager(database=self.data.db),
             bitcoin_cash_manager=BitcoinCashManager(database=self.data.db),
             solana_manager=SolanaManager(
-                greenlet_manager=self.greenlet_manager,
-                database=self.data.db,
+                node_inquirer=SolanaInquirer(
+                    greenlet_manager=self.greenlet_manager,
+                    database=self.data.db,
+                ),
+                premium=self.premium,
             ),
             msg_aggregator=self.msg_aggregator,
             database=self.data.db,
